Drop the unused full-view navigation helper, share the drawer test mocks and shorten the changeset
