+++ .changeset/optimization-dedicated-view.md
@@ -2,4 +2,4 @@
 "@hashintel/petrinaut": patch
 ---
 
-Add a full-size presentation of one optimization beside the drawer (Open full view, Back to list, Show in drawer), an Objective by step chart with the best so far as a line, a header line naming the best step so far with a Still improving / Converging verdict while the study runs, and results charts once a study is over. The summary strip's Best stat is labelled "Best step so far".
+Add a full view for optimization studies with an Objective by step chart and a header naming the best step so far.
