+++ libs/@hashintel/petrinaut/src/ui/views/Editor/panels/SimulateView/optimizations/optimization-full-view.test.tsx
@@ -37,38 +37,17 @@ vi.mock("./optimization-surface", () => ({
 }));
 
 // uPlot cannot mount in jsdom; the cards around the charts are real.
-vi.mock("./study-view/objective-history-chart", async () => {
-  const chartCard = await vi.importActual<
-    typeof import("../shared/chart-card")
-  >("../shared/chart-card");
-  return {
-    ObjectiveHistoryCard: ({ plotHeight }: { plotHeight: number }) => (
-      <chartCard.ChartCard title="Objective by step" bodyHeight={plotHeight}>
-        <div data-testid="objective-history" />
-      </chartCard.ChartCard>
-    ),
-  };
-});
+vi.mock("./study-view/objective-history-chart", () =>
+  import("../shared/metric-timeline-test-stubs").then((stubs) =>
+    stubs.mockObjectiveHistoryCardModule(),
+  ),
+);
 
-vi.mock("../experiments/experiment-metric-timeline", async () => {
-  const [menu, describeView, viewState] = await Promise.all([
-    vi.importActual<
-      typeof import("../experiments/experiment-metric-timeline/metric-view-menu")
-    >("../experiments/experiment-metric-timeline/metric-view-menu"),
-    vi.importActual<
-      typeof import("../experiments/experiment-metric-timeline/describe-metric-view")
-    >("../experiments/experiment-metric-timeline/describe-metric-view"),
-    vi.importActual<
-      typeof import("../experiments/experiment-metric-timeline/view-state")
-    >("../experiments/experiment-metric-timeline/view-state"),
-  ]);
-  return {
-    MetricViewMenu: menu.MetricViewMenu,
-    describeMetricView: describeView.describeMetricView,
-    DEFAULT_METRIC_VIEW_SETTINGS: viewState.DEFAULT_METRIC_VIEW_SETTINGS,
-    ExperimentMetricTimeline: () => <div data-testid="metric-timeline" />,
-  };
-});
+vi.mock("../experiments/experiment-metric-timeline", () =>
+  import("../shared/metric-timeline-test-stubs").then((stubs) =>
+    stubs.mockExperimentMetricTimelineModule(),
+  ),
+);
 
 afterEach(cleanup);
